chore: learn vocab card own component
