+++ src/screens/Learn/Query/Query.scss
@@ -26,88 +26,6 @@
     display: flex;
     justify-content: center;
     align-items: center;
-
-    .card {
-      background-color: transparent;
-      width: 100%;
-      height: 100%;
-      display: flex;
-      perspective: 1000px;
-
-      .card-inner {
-        position: relative;
-        width: 50%;
-        min-width: 450px;
-        max-width: 600px;
-        height: 70%;
-        min-height: 400px;
-        margin: auto;
-        text-align: center;
-        transition: transform 0.8s;
-        transform-style: preserve-3d;
-
-        &.flipped {
-          transform: rotateY(180deg);
-        }
-      }
-
-      .card-front,
-      .card-back {
-        position: absolute;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        width: 100%;
-        height: 100%;
-        backface-visibility: hidden;
-        border-radius: 10px;
-        box-shadow: 1px 3px 5px $color-box-shadow;
-        cursor: pointer;
-      }
-
-      .card-front {
-        color: $color-main-text-inverse;
-        background-color: $color-background-inverse;
-
-        .card-front-inner {
-          padding: 20px;
-          height: 100%;
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-        }
-      }
-
-      .card-back {
-        transform: rotateY(180deg);
-        color: $color-main-text-inverse;
-        background-color: $color-primary-dark;
-
-        .card-back-inner {
-          padding: 20px;
-          height: 100%;
-          display: flex;
-          flex-direction: column;
-          justify-content: space-between;
-        }
-      }
-
-      .my-20 {
-        margin: 20px 0;
-      }
-
-      .translations {
-        border: 1px solid $color-primary;
-        border-radius: 10px;
-        padding: 10px;
-        font-size: 24px;
-      }
-
-      .continue {
-        display: flex;
-        justify-content: space-evenly;
-      }
-    }
   }
 
   .footer {
